+++ package.json
@@ -4,7 +4,7 @@
   "description": "Management global state with out provider",
   "main": "index.cjs.js",
   "module": "index.js",
-  "types": "tempTypes",
+  "types": "types",
   "repository": "https://github.com/jungpaeng/react-manage-global-state.git",
   "author": "jungpaeng <meis1541@naver.com> (https://github.com/jungpaeng)",
   "license": "MIT",
