+++ src/index.ts
@@ -1,8 +1,3 @@
 // lib
 export * from './store';
 export * from './middleware';
-
-// types
-export * from './types/persist';
-export * from './types/redux';
-export * from './types/store';
