+++ src/store/useGlobalStore.ts
@@ -1,5 +1,7 @@
 import React from 'react';
-import { CreateStoreReturnValue, EqualityFunction } from '../types/store';
+import { CreateStoreReturnValue } from './createStore';
+
+export type EqualityFunction<T> = (prev: T, next: T) => boolean;
 
 const defaultEqualityFn = (prev: any, next: any) => prev === next;
