--- src/types/redux.ts
@@ -1,7 +1,0 @@
-export interface ReducerAction {
-  type: string;
-  [key: string]: any;
-}
-
-export type DispatchType = (action: ReducerAction) => void;
-export type ReducerFunction<T> = (state: T, action: ReducerAction) => T;
