+++ tsconfig.json
@@ -5,7 +5,7 @@
     "jsx": "preserve",
     "allowSyntheticDefaultImports": true,
     "declaration": true,
-    "declarationDir": "dist/tempTypes",
+    "declarationDir": "dist/types",
     "esModuleInterop": true,
     "moduleResolution": "node",
     "skipLibCheck": true
